Scrape US Census Data

Takes in a city/state and outputs demographic data from the Missouri Census Data Center

getCensus(cities, radius)

Arguments

cities
list of cities to be passed

radius
numeric number used to create a boundary around the city coordinates

Value

Data frame containing demographic data

Examples

radius <- 2
cities <- c("Boston", "Los Angeles")
df <- getCensus(cities, radius)

#> Information from URL : http://maps.googleapis.com/maps/api/geocode/json?address=Boston&sensor=false

#> Information from URL : http://maps.googleapis.com/maps/api/geocode/json?address=Los%20Angeles&sensor=false

#> [[1]]
#> [1] "http://mcdc.missouri.edu/cgi-bin/broker?_PROGRAM=websas.caps10acsb.sas&SERVICE=appdev&sitename=Boston&longitude=-71.0588801&latitude=42.3600825&radii=2&dprofile=on&eprofile=on&sprofile=on&hprofile=on&cntypops=on&printdetail=on&_debug="
#> 
#> [[2]]
#> [1] "http://mcdc.missouri.edu/cgi-bin/broker?_PROGRAM=websas.caps10acsb.sas&SERVICE=appdev&sitename=LosAngeles&longitude=-118.2436849&latitude=34.0522342&radii=2&dprofile=on&eprofile=on&sprofile=on&hprofile=on&cntypops=on&printdetail=on&_debug="
#> 

head(df)

#>     sitename radius    Longitude   Latitude TotPopSF1 AreaSQMI LandSQMI
#> 1     Boston      2  -71.0588801 42.3600825   175,111 10.93563     8.74
#>    IntPtLat    IntPtLon vintage period  TotPop Age0_4 pctAge0_4 Age5_9
#> 1 42.358199  -71.063640    2015    5yr 188,658  9,654       5.1  5,962
#>   pctAge5_9 Age10_14 pctAge10_14 Age15_19 pctAge15_19 Age20_24 pctAge20_24
#> 1       3.2    5,020         2.7   10,070         5.3   19,258        10.2
#>   Age25_34 pctAge25_34 Age35_44 pctAge35_44 Age45_54 pctAge45_54 Age55_59
#> 1   55,847        29.6   25,497        13.5   19,754        10.5    9,238
#>   pctAge55_59 Age60_64 pctAge60_64 Age65_74 pctAge65_74 Age75_84 pctAge75_84
#> 1         4.9    7,950         4.2   11,427         6.1    6,247         3.3
#>   Over85 pctOver85 MedianAge  Over15 pctOver15 Under18 pctUnder18  Over18
#> 1  2,734       1.4      33.5 168,022      89.1  23,492       12.5 165,166
#>   pctOver18  Over21 pctOver21  Over25 pctOver25 Over62 pctOver62 Over65
#> 1      87.5 154,034      81.6 138,694      73.5 25,320      13.4 20,408
#>   pctOver65  Males pctMales Over65Males pctOver65Males Females pctFemales
#> 1      10.8 94,297     50.0       9,093            9.6  94,361       50.0
#>   Over65Females pctOver65Females OneRace pctOneRace  White1 pctWhite1 Black1
#> 1        11,315             12.0 175,446       93.0 131,740      69.8 12,875
#>   pctBlack1 Indian1 pctIndian1 Asian1 pctAsian1 HawnPI1 pctHawnPI1 Other1
#> 1       6.8     541        0.3 23,291      12.3      24        0.0  6,975
#>   pctOther1 MultiRace pctMultiRace  White2 pctWhite2 Black2 pctBlack2 Indian2
#> 1       3.7    13,212          7.0 143,961      76.3 22,538      11.9   1,450
#>   pctIndian2 Asian2 pctAsian2 HawnPI2 pctHawnPI2 Other2 pctOther2 HispanicPop
#> 1        0.8 25,423      13.5     590        0.3  8,445       4.5      35,320
#>   pctHispanicPop NonHispPop pctNonHispPop NonHispWhite pctNonHispWhite
#> 1           18.7    153,338          81.3      113,130            60.0
#>   NonHispBlack pctNonHispBlack NonHispAmInd pctNonHispAmInd NonHispAsian
#> 1       11,508             6.1          195             0.1       23,174
#>   pctNonHispAsian NonHispHawnPI pctNonHispHawnPI TotHHs HHInc0 pctHHInc0
#> 1            12.3            24              0.0 87,490  9,267      10.6
#>   HHInc10 pctHHInc10 HHInc15 pctHHInc15 HHInc25 pctHHInc25 HHInc35 pctHHInc35
#> 1   5,117        5.8   6,631        7.6   5,233        6.0   7,141        8.2
#>   HHInc50 pctHHInc50 HHInc75 pctHHInc75 HHInc100 pctHHInc100 HHInc150
#> 1  11,155       12.8   8,746       10.0   13,730        15.7    7,565
#>   pctHHInc150 HHInc200 pctHHInc200 NumHHEarnings pctNumHHEarnings NumHHSocSec
#> 1         8.6   12,905        14.8        71,329             81.5      14,817
#>   pctNumHHSocSec NumHHRetInc pctNumHHRetInc NumHHSuppSecInc pctNumHHSuppSecInc
#> 1           16.9       5,959            6.8           4,945                5.7
#>   NumHHPubAssist pctNumHHPubAssist MedianHHInc AvgHHInc AvgHHEarnings
#> 1          2,259               2.6     $83,486 $116,894      $124,863
#>   AvgHHSocSec AvgHHRetInc AvgHHSuppSecInc AvgHHPubAssist FamHHs pctFamHHs
#> 1     $15,865     $31,835          $8,965         $3,620 33,192      37.9
#>   FamHHInc0 pctFamHHInc0 FamHHInc10 pctFamHHInc10 FamHHInc15 pctFamHHInc15
#> 1     1,965          5.9      1,176           3.5      2,829           8.5
#>   FamHHInc25 pctFamHHInc25 FamHHInc35 pctFamHHInc35 FamHHInc50 pctFamHHInc50
#> 1      2,419           7.3      3,029           9.1      3,388          10.2
#>   FamHHInc75 pctFamHHInc75 FamHHInc100 pctFamHHInc100 FamHHInc150
#> 1      2,903           8.7       4,845           14.6       3,186
#>   pctFamHHInc150 FamHHInc200 pctFamHHInc200 AvgFamInc     PCI NonFamHHs
#> 1            9.6       7,452           22.4  $151,293 $56,254    54,298
#>   pctNonFamHHs MedianNonFamInc AvgNonFamInc PovUniverseFamilyPop PoorInFamily
#> 1         62.1         $74,360      $94,348               95,288       13,677
#>   pctPoorInFamily PoorFamilies pctPoorFamilies PovRatioUnderHalf
#> 1            14.4        4,254            12.8            16,664
#>   PovRatiov5tov99 PovRatio1to2 PovRatioOver2  Over16 LaborForce pctLaborForce
#> 1          16,922       26,595       118,181 167,276    120,047         10959
#>   CivLabForce pctCivLabForce EmployedCLF pctEmployedCLF UnemployedCLF
#> 1     119,791          10936     112,422          14277         7,369
#>   pctUnemployedCLF Military pctMilitary NotInLF pctNotInLF OwnKidsUnder6
#> 1             1023      256        22.7  47,229       4341        10,653
#>   OwnKidsU6AllPrntsWk pctOwnKidsU6AllPrntsWk OwnKidsOver6 OwnKidsO6AllPrntsWk
#> 1               7,452                   70.0       11,814               8,437
#>   pctOwnKidsO6AllPrntsWk Worker16 Commuters pctCommuters DriveAlone
#> 1                   71.4  110,448   105,569         95.6     29,192
#>   pctDriveAlone Carpool pctCarpool PublicTrans pctPublicTrans WalkToWork
#> 1          26.4   4,979        4.5      35,938           32.5     31,658
#>   pctWalkToWork OtherCommute pctOtherCommute WorkAtHome pctWorkAtHome
#> 1          28.7        3,802             3.4      4,879           4.4
#>   ManProfOccs pctManProfOccs ServiceOccs pctServiceOccs SalesOffOccs
#> 1      66,381           59.0      18,587           16.5       19,296
#>   pctSalesOffOccs FarmFishOccs pctFarmFishOccs ConsOccs pctConsOccs TransOccs
#> 1            17.2          175             0.2    2,863         2.5     5,120
#>   pctTransOccs Agriculture pctAgriculture Construction pctConstruction
#> 1          4.6         166            0.1        2,774             2.5
#>   Manufacturing pctManufacturing WholesaleTrade pctWholesaleTrade RetailTrade
#> 1         5,837              5.2          2,096               1.9       6,406
#>   pctRetailTrade Transportation pctTransportation Information pctInformation
#> 1            5.7          2,361               2.1       3,187            2.8
#>   Finance_Ins pctFinance_Ins Professional pctProfessional Educ_Health_SocSvcs
#> 1      15,602           13.9       25,310            22.5              28,484
#>   pctEduc_Health_SocSvcs RecreationEtc pctRecreationEtc OtherIndustries
#> 1                   25.3        13,168             11.7           3,850
#>   pctOtherIndustries PublicAdmin pctPublicAdmin PrivWageWorkers
#> 1                3.4       3,181            2.8          99,911
#>   pctPrivWageWorkers GovWorkers pctGovWorkers SelfEmpWorkers pctSelfEmpWorkers
#> 1               88.9      7,737           6.9          4,665               4.1
#>   UnpaidFamWorkers pctUnpaidFamWorkers Families pctFamilies FamsWithKids
#> 1              109                 0.1   33,192        37.9       12,322
#>   pctFamsWithKids MarriedCouples pctMarriedCouples CouplesWithKids
#> 1            14.1         23,174              26.5           7,431
#>   pctCouplesWithKids SingleMaleFamilies pctSingleMaleFamilies SingleFathers
#> 1                8.5              2,731                   3.1           985
#>   pctSingleFathers SingleFemaleFamilies pctSingleFemaleFamilies SingleMothers
#> 1              1.1                7,287                     8.3         3,906
#>   pctSingleMothers HHsWithKids pctHHsWithKids HHsWithElders pctHHsWithElders
#> 1              4.5      13,424           15.3        15,859             18.1
#>   AvgHHSize AvgFamSize   HHPop pctHHPop FamHHpop pctFamHHpop NonFamHHpop
#> 1      2.02       2.87 176,397     93.5   98,939        8307      77,458
#>   pctNonFamHHpop GrpQuarters pctGrpQuarters Householder pctHouseholder Spouse
#> 1           6893      12,261            6.5      87,490           7759 23,314
#>   pctSpouse  Child pctChild OtherRelative pctOtherRelative NonRelative
#> 1      2075 29,254     2319         9,528            782.2      26,811
#>   pctNonRelative UnmarriedPartner pctUnmarriedPartner UMPartnerHHsPerK
#> 1           2266            7,111               598.4               81
#>   NeverMarried pctNeverMarried Married pctMarried Separated pctSeparated
#> 1       92,560            55.1  54,307       32.3     3,492          2.1
#>   Widowed pctWidowed Divorced pctDivorced   Over3 EnrolledOver3
#> 1   5,926        3.5   11,737         7.0 182,507        45,849
#>   pctEnrolledOver3 InNursery pctInNursery InKindergarten pctInKindergarten
#> 1             25.1     2,387          5.2          1,614               3.5
#>   InElementary pctInElementary InHighSchool pctInHighSchool InCollege
#> 1        8,584            18.7        3,558             7.8    29,706
#>   pctInCollege LessThan9th pctLessThan9th SomeHighSchool pctSomeHighSchool
#> 1         64.8      11,450            8.3          7,202               5.2
#>   HighSchool pctHighSchool SomeCollege pctSomeCollege Associates pctAssociates
#> 1     19,271          13.9      12,113            8.7      3,807           2.7
#>   Bachelors pctBachelors GradProf pctGradProf HighSchoolOrMore
#> 1    42,386         30.6   42,465        30.6          120,042
#>   pctHighSchoolOrMore Bachelorsormore pctBachelorsormore Civilian pctCivilian
#> 1                86.6          84,851               61.2  164,910       157.9
#>   Veteran pctVeteran TotHUs OccHUs pctOccHUs OwnerOcc pctOwnerOcc RenterOcc
#> 1   5,301        3.2 96,529 87,490      90.6   28,716        32.8    58,774
#>   pctRenterOcc AvgOwnerHHSize AvgRenterHHSize VacHUs pctVacHUs VacantForSale
#> 1         67.2           2.01            2.02  9,039       9.4           494
#>   pctVacantForSale VacantForRent pctVacantForRent VacantSeasonal
#> 1              5.5         1,720             19.0          2,759
#>   pctVacantSeasonal TotalOwnerUnits OwnerVacRate TotalRentalUnits RenterVacRate
#> 1              30.5          29,210          1.7           60,494           2.8
#>   PersonsInOwnerUnits pctPersonsInOwnerUnits PersonsInRenterUnits
#> 1              57,764                   30.6              118,633
#>   pctPersonsInRenterUnits Units1 pctUnits1 Units1Detached pctUnits1Detached
#> 1                    62.9  7,812       8.1          2,539              32.5
#>   Units1Attached pctUnits1Attached Units2 pctUnits2 Units3_4 pctUnits3_4
#> 1          5,273              67.5  6,434       6.7   22,979        23.8
#>   Units5_9 pctUnits5_9 Units10_19 pctUnits10_19 Units20up pctUnits20up
#> 1   16,159        16.7      8,072           8.4    34,895         36.1
#>   MobileHomes pctMobileHomes BoatRV pctBoatRV MobileHomesPerK Built2010orLater
#> 1          77            0.1    101       0.1               1            1,180
#>   pctBuilt2010orLater Built2000_2009 pctBuilt2000_2009 Built1990_1999
#> 1                 1.2         10,499              10.9          3,637
#>   pctBuilt1990_1999 Built1980_1989 pctBuilt1980_1989 Built1970_1979
#> 1               3.8          6,562               6.8          8,069
#>   pctBuilt1970_1979 Built1960_1969 pctBuilt1960_1969 Built1950_1959
#> 1               8.4          5,557               5.8          3,979
#>   pctBuilt1950_1959 Built1940_1949 pctBuilt1940_1949 BuiltBefore1940
#> 1               4.1          3,585               3.7          53,461
#>   pctBuiltBefore1940 MovedIn2010orLater pctMovedIn2010orLater MovedIn2000_2009
#> 1               55.4             43,019                  49.2           29,775
#>   pctMovedIn2000_2009 MovedIn1990_1999 pctMovedIn1990_1999 MovedIn1980_1989
#> 1                34.0            8,043                 9.2            3,274
#>   pctMovedIn1980_1989 MovedInBefore1980 pctMovedInBefore1980 NoVehicles
#> 1                 3.7             3,379                  3.9     38,283
#>   pctNoVehicles Vehicles1 pctVehicles1 Vehicles2 pctVehicles2 VehiclesGE3
#> 1          43.8    36,817         42.1    10,798         12.3       1,592
#>   pctVehiclesGE3 HHFUtilGas pctHHFUtilGas HHFLPGas pctHHFLPGas HHFElectric
#> 1            1.8     43,030          49.2    1,624         1.9      34,191
#>   pctHHFElectric HHFKerosene pctHHFKerosene HHFCoal pctHHFCoal HHFWood
#> 1           39.1       6,850            7.8      19        0.0      67
#>   pctHHFWood HHFSolar pctHHFSolar HHFOther pctHHFOther HHFNoFuel pctHHFNoFuel
#> 1        0.1       51         0.1      640         0.7     1,018          1.2
#>   NoPlumbing pctNoPlumbing NoKitchen pctNoKitchen NoPhone pctNoPhone
#> 1        327           0.4       601          0.7   2,357        2.7
#>   PersonsPerRoomLow pctPersonsPerRoomLow PersonsPerRoomMedium
#> 1            84,291                 96.3                1,862
#>   pctPersonsPerRoomMedium PersonsPerRoomHigh pctPersonsPerRoomHigh HvalUnder50
#> 1                     2.1              1,337                   1.5         367
#>   pctHvalUnder50 Hval50 pctHval50 Hval100 pctHval100 Hval150 pctHval150 Hval200
#> 1            1.3     80       0.3     355        1.2     485        1.7   2,325
#>   pctHval200 Hval300 pctHval300 Hval500 pctHval500 HvalOverMillion
#> 1        8.1   9,402       32.7  10,356       36.1           3,503
#>   pctHvalOverMillion HvalOver2Million pctHvalOver2Million MedianHValue
#> 1               12.2            1,843                 6.4     $623,962
#>   AvgHValue HUsMort pctHUsMort HUsMortOver30Pct pctHUsMortOver30Pct
#> 1  $813,091  21,155       73.7            6,910                32.7
#>   MedianOwnerCostsMort HUsNoMort pctHUsNoMort HUsNoMortOver30Pct
#> 1               $2,704     7,561         26.3              1,872
#>   pctHUsNoMortOver30Pct MedianOwnerCostsNoMort CashRenter pctCashRenter
#> 1                  24.8                 $1,143     57,602          98.0
#>   NoCashRenter pctNoCashRenter MedianGrossRent AvgGrossRent CashRenterOver30Pct
#> 1        1,172             2.0          $1,582       $1,621              26,339
#>   pctCashRenterOver30Pct CashRenterOver750 pctCashRenterOver750 Persons_Sampled
#> 1                   44.8            34,927                 59.4           15277
#>   PctPersons_Sampled HUs_Sampled PctHUs_Sampled PctOver5 PctOver18Males
#> 1                8.1        7458            7.7     94.9           87.3
#>   PctOver18Females PovUniverse  poor Pctpoor PctPovRatioUnderHalf
#> 1             87.8           0 33586    18.8                  9.3
#>   PctPovRatiov5tov99 PctPovRatio1to2 PctpovratioOver2
#> 1                9.5            14.9             66.3
#>  [ reached getOption("max.print") -- omitted 1 row ]